Making clippy happy and modernizing rust edition

+++ src/internal/iterator.rs
@@ -15,7 +15,6 @@
  * file, You can obtain one at <https://mozilla.org/MPL/2.0/>.
  */
 
-
 //! Iterators for matches in the RPM database
 
 use super::{header::Header, tag::Tag, ts::GlobalTS};
@@ -56,7 +55,7 @@ impl MatchIterator {
                         txn.as_mut_ptr(),
                         tag as librpm_sys::rpm_tag_t,
                         key.as_ptr() as *const c_void,
-                        key.len() as u64,
+                        key.len(),
                     )
                 };
